NO JIRA fix code smells after moving to lang3 library

+++ server/sonar-webserver-webapi/src/main/java/org/sonar/server/source/ws/ScmAction.java
@@ -98,7 +98,7 @@ public class ScmAction implements SourcesWsAction {
   public void handle(Request request, Response response) {
     String fileKey = request.mandatoryParam("key");
     int from = Math.max(request.mandatoryParamAsInt("from"), 1);
-    int to = (Integer) ObjectUtils.defaultIfNull(request.paramAsInt("to"), Integer.MAX_VALUE);
+    int to = ObjectUtils.defaultIfNull(request.paramAsInt("to"), Integer.MAX_VALUE);
     boolean commitsByLine = request.mandatoryParamAsBoolean("commits_by_line");
 
     try (DbSession dbSession = dbClient.openSession(false)) {
